HW3 - Block Cipher/requirements.txt
numpy
opencv-python
pycryptodome
argparse